js 파일들 컨벤션 통일
