+++ src/public/js/user/close-modal.js
