+++ src/views/user/user.ejs
@@ -205,13 +205,14 @@
   <script src="/js/sb-admin-2.min.js"></script>
 
   <!-- 모달 닫을시에 입력값 초기화 -->
-  <script src="/js/user/closeModal.js"></script>
+  <script src="/js/user/close-modal.js"></script>
 
   <!-- Send Ajax 메소드 추가 -->
   <script src="/js/common/ajax.js"></script>
   
   <!-- 유저 생성, 유저 상세조회에 대한 이벤트 -->
-  <script type="module" src="/js/user/user.js"></script>
+  <script type="module" src="/js/user/user-search.js"></script>
+  <script type="module" src="/js/user/user-create.js"></script>
 
 </body>
