+++ src/views/user/userDetail.ejs
@@ -337,10 +337,10 @@
   <script src="/js/common/ajax.js"></script>
 
   <!-- 유저정보 수정관련 이벤트 처리-->
-  <script type="module" src="/js/user/userModify.js"></script>
+  <script type="module" src="/js/user/user-modify.js"></script>
 
   <!-- 유저 키워드 추가 모달 이벤트 처리-->
-  <script type="module" src="/js/user/userKeyword.js"></script>
+  <script type="module" src="/js/user/user-keyword-search.js"></script>
   
   <!-- 토탈 키워드 추가 모달-->
   <%- include ('./userTotalKeywordModal.ejs') %>
